+++ Object/Player.cpp
@@ -274,8 +274,10 @@ void Player::Draw()
 	if(!m_isHealthBer)
 	{
 		SetDrawBlendMode(DX_BLENDMODE_ALPHA, 230);
-		DrawBox(m_pos.x - 35, m_pos.y - 30, m_pos.x - 35 + 75, m_pos.y - 30 + 10, 0xff0000, true);
-		DrawExtendGraph(m_pos.x - 35, m_pos.y - 30, m_playerHealthSizeX + m_pos.x - 35 + 75, m_playerHealthSizeY + m_pos.y - 30 + 10,m_hHealthBer,true);
+		DrawBox(static_cast<int>(m_pos.x) - 35, static_cast<int>(m_pos.y) - 30,
+			static_cast<int>(m_pos.x) - 35 + 75, static_cast<int>(m_pos.y) - 30 + 10, 0xff0000, true);
+		DrawExtendGraph(static_cast<int>(m_pos.x) - 35, static_cast<int>(m_pos.y) - 30,
+			m_playerHealthSizeX + m_pos.x - 35 + 75, m_playerHealthSizeY + m_pos.y - 30 + 10,m_hHealthBer,true);
 		DrawBox(m_pos.x - 35, m_pos.y - 30, m_pos.x - 35 + 75, m_pos.y - 30 + 10, 0xffffff, false);
 		SetDrawBlendMode(DX_BLENDMODE_NOBLEND, 0);
 	}
@@ -285,7 +287,8 @@ void Player::Draw()
 	SetDrawBlendMode(DX_BLENDMODE_ALPHA, 64);
 	for (int i = 0; i < 4 ; i++)//”wŒiƒ‰ƒCƒg
 	{
-		DrawRotaGraph(m_pos.x - 3 + GetRand(5), m_pos.y + 20 + GetRand(5), 3, 0, m_hPlayerLighting, true, false);
+		DrawRotaGraph(static_cast<int>(m_pos.x) - 3 + GetRand(5), static_cast<int>(m_pos.y) + 20 + GetRand(5)
+			, 3, 0, m_hPlayerLighting, true, false);
 	}
 	SetDrawBlendMode(DX_BLENDMODE_NOBLEND, 0);
 	if (m_isIdleMove && !m_isAttackMove && !m_isDamageMove && !m_isJumpMove && !m_isCrouchingMove)//‰½‚à‚µ‚È‚¢
